+++ srunner/scenarios/master_scenario.py
@@ -81,9 +81,8 @@ class MasterScenario(BasicScenario):
         collision_criterion = CollisionTest(self.ego_vehicles[0], terminate_on_failure=False)
 
         route_criterion = InRouteTest(self.ego_vehicles[0],
-                                      radius=30.0,
                                       route=route,
-                                      offroad_max=20,
+                                      offroad_max=30,
                                       terminate_on_failure=True)
 
         completion_criterion = RouteCompletionTest(self.ego_vehicles[0], route=route)