+++ Docs/CHANGELOG.md
@@ -12,6 +12,7 @@
 * New Features:
     - Added new OutsideRouteLanesTest atomic criter that encompasses both SidewalkTest and WrongLaneTest, returning
     the percentage of route that has been traveled outside the lane.
+    - InRouteTest is now more forgiving. The max distance has been increased, but staying above the previous one will eventually also cause failure
     - Changed SidewalkTest atomic criteria to also track other type of out of lane conditions
     - SidewalkTest and WrongLaneTest atomic criterias now track the amount of meters traversed.
     - CollisionTest atomic criteria now correctly ignores multiple micro-collisions with the same object